Trait mirrord_config::config::MirrordConfig

pub trait MirrordConfig {
    type Generated;

    // Required method
    fn generate_config(
        self,
        context: &mut ConfigContext
    ) -> Result<Self::Generated>;
}

Main configuration creation trait of mirrord-config

Required Associated Types

type Generated

The resulting struct you plan on using in the rest of your code

Required Methods

fn generate_config(self, context: &mut ConfigContext) -> Result<Self::Generated>

Load configuration from all sources and output as Self::Generated Pass reference to list of warnings which callee can add warnings into.