Struct mirrord_config::LayerConfig

pub struct LayerConfig {

    pub accept_invalid_certificates: bool,
    pub skip_processes: Option<VecOrSingle<String>>,
    pub pause: bool,
    pub connect_tcp: Option<String>,
    pub connect_agent_name: Option<String>,
    pub connect_agent_port: Option<u16>,
    pub operator: bool,
    pub kubeconfig: Option<String>,
    pub sip_binaries: Option<VecOrSingle<String>>,
    pub target: TargetConfig,
    pub agent: AgentConfig,
    pub feature: FeatureConfig,
    pub telemetry: bool,

}

mirrord allows for a high degree of customization when it comes to which features you want to enable, and how they should function.

All of the configuration fields have a default value, so a minimal configuration would be no configuration at all.

To help you get started, here are examples of a basic configuration file, and a complete configuration file containing all fields.

Basic config.json {#root-basic}

{
  "target": "pod/bear-pod",
  "feature": {
    "env": true,
    "fs": "read",
    "network": true
  }
}

Complete config.json {#root-complete}

{
  "accept_invalid_certificates": false,
  "skip_processes": "ide-debugger",
  "pause": false,
  "target": {
    "path": "pod/bear-pod",
    "namespace": "default"
  },
  "connect_tcp": null,
  "agent": {
    "log_level": "info",
    "namespace": "default",
    "image": "ghcr.io/metalbear-co/mirrord:latest",
    "image_pull_policy": "IfNotPresent",
    "image_pull_secrets": [ { "secret-key": "secret" } ],
    "ttl": 30,
    "ephemeral": false,
    "communication_timeout": 30,
    "startup_timeout": 360,
    "network_interface": "eth0",
    "flush_connections": true
  },
  "feature": {
    "env": {
      "include": "DATABASE_USER;PUBLIC_ENV",
      "exclude": "DATABASE_PASSWORD;SECRET_ENV",
      "overrides": {
        "DATABASE_CONNECTION": "db://localhost:7777/my-db",
        "LOCAL_BEAR": "panda"
      }
    },
    "fs": {
      "mode": "write",
      "read_write": ".+\.json" ,
      "read_only": [ ".+\.yaml", ".+important-file\.txt" ],
      "local": [ ".+\.js", ".+\.mjs" ]
    },
    "network": {
      "incoming": {
        "mode": "steal",
        "http_header_filter": {
          "filter": "host: api\..+",
          "ports": [80, 8080]
        },
        "port_mapping": [[ 7777, 8888 ]],
        "ignore_localhost": false,
        "ignore_ports": [9999, 10000]
      },
      "outgoing": {
        "tcp": true,
        "udp": true,
        "ignore_localhost": false,
        "unix_streams": "bear.+"
      },
      "dns": false
    },
    "capture_error_trace": false
  },
  "operator": true,
  "kubeconfig": "~/.kube/config",
  "sip_binaries": "bash"
  "telemetry": true,
}

Options {#root-options}

Fields
accept_invalid_certificates: bool

accept_invalid_certificates {#root-accept_invalid_certificates}

Controls whether or not mirrord accepts invalid TLS certificates (e.g. self-signed certificates).

Defaults to false.

skip_processes: Option<VecOrSingle<String>>

skip_processes {#root-skip_processes}

Allows mirrord to skip unwanted processes.

Useful when process A spawns process B, and the user wants mirrord to operate only on process B. Accepts a single value, or multiple values separated by ;.

 {
  "skip_processes": "bash;node"
 }

pause: bool

pause {#root-pause}

Controls target pause feature. Unstable.

With this feature enabled, the remote container is paused while this layer is connected to the agent.

Defaults to false.

connect_tcp: Option<String>

connect_tcp {#root-connect_tpc}

IP:PORT to connect to instead of using k8s api, for testing purposes.

{
  "connect_tcp": "10.10.0.100:7777"
}

connect_agent_name: Option<String>

connect_agent_name {#root-connect_agent_name}

Agent name that already exists that we can connect to.

Keep in mind that the intention here is to allow reusing a long living mirrord-agent pod, and not to connect multiple (simultaneos) mirrord instances to a single mirrord-agent, as the later is not properly supported without the use of mirrord-operator.

{
  "connect_agent_name": "mirrord-agent-still-alive"
}

connect_agent_port: Option<u16>

connect_agent_port {#root-connect_agent_port}

Agent listen port that already exists that we can connect to.

{
  "connect_agent_port": "8888"
}

operator: bool

operator {#root-operator}

Allow to lookup if operator is installed on cluster and use it.

Defaults to true.

kubeconfig: Option<String>

kubeconfig {#root-kubeconfig}

Path to a kubeconfig file, if not specified, will use KUBECONFIG, or ~/.kube/config, or the in-cluster config.

{
 "kubeconfig": "~/bear/kube-config"
}

sip_binaries: Option<VecOrSingle<String>>

sip_binaries {#root-sip_binaries}

Binaries to patch (macOS SIP).

Use this when mirrord isn’t loaded to protected binaries that weren’t automatically patched.

Runs endswith on the binary path (so bash would apply to any binary ending with bash while /usr/bin/bash would apply only for that binary).

{
 "sip_binaries": "bash;python"
}

target: TargetConfig

target {#root-target}

agent: AgentConfig

agent {#root-agent}

feature: FeatureConfig

feature {#root-feature}

telemetry: bool

telemetry {#root-telemetry}

Controls whether or not mirrord sends telemetry data to MetalBear cloud. Telemetry sent doesn’t contain personal identifiers or any data that should be considered sensitive. It is used to improve the product. For more information

Implementations

impl LayerConfig

pub fn from_env() -> Result<Self, ConfigError>

pub fn verify(&self) -> Result<(), ConfigError>

Verify that there are no conflicting settings. We don’t call it from from_env since we want to verify it only once (from cli)