Struct mirrord_config::LayerFileConfig

pub struct LayerFileConfig {
    pub accept_invalid_certificates: Option<bool>,
    pub skip_processes: Option<VecOrSingle<String>>,
    pub target: Option<<TargetConfig as FromMirrordConfig>::Generator>,
    pub connect_tcp: Option<String>,
    pub connect_agent_name: Option<String>,
    pub connect_agent_port: Option<u16>,
    pub agent: Option<<AgentConfig as FromMirrordConfig>::Generator>,
    pub feature: Option<<FeatureConfig as FromMirrordConfig>::Generator>,
    pub operator: Option<bool>,
    pub kubeconfig: Option<String>,
    pub sip_binaries: Option<VecOrSingle<String>>,
}

Main struct for mirrord-layer’s configuration

Examples

Run mirrord with read-only file operations, mirroring traffic, skipping unwanted processes:

# mirrord-config.toml

target = "pod/sample-pod-1234"
skip_processes = ["ide-debugger", "ide-service"] # we don't want mirrord to hook into these

[agent]
log_level = "debug"
ttl = 1024 # seconds

[feature]
fs = "read" # default

[feature.network]
incoming = "mirror" # default

Run mirrord with read-write file operations, stealing traffic, accept local TLS certificates, use a custom mirrord-agent image:

# mirrord-config.toml

target = "pod/sample-pod-1234"
accept_invalid_certificates = true

[agent]
log_level = "trace"
ttl = 1024 # seconds
image = "registry/mirrord-agent-custom:latest"
image_pull_policy = "Always"

[feature]
fs = "write"

[feature.network]
incoming = "steal"

Fields
accept_invalid_certificates: Option<bool>

Controls whether or not mirrord accepts invalid TLS certificates (e.g. self-signed certificates).

skip_processes: Option<VecOrSingle<String>>

Allows mirrord to skip unwanted processes.

Useful when process A spawns process B, and the user wants mirrord to operate only on process B.

target: Option<<TargetConfig as FromMirrordConfig>::Generator>

Specifies the running pod to mirror.

Supports:

pod/{sample-pod}/[container]/{sample-container};

podname/{sample-pod}/[container]/{sample-container};

deployment/{sample-deployment}/[container]/{sample-container};

connect_tcp: Option<String>

IP:PORT to connect to instead of using k8s api, for testing purposes.

connect_agent_name: Option<String>

Agent name that already exists that we can connect to.

connect_agent_port: Option<u16>

Agent listen port that already exists that we can connect to.

agent: Option<<AgentConfig as FromMirrordConfig>::Generator>

Agent configuration, see agent::AgentFileConfig.

feature: Option<<FeatureConfig as FromMirrordConfig>::Generator>

Controls mirrord features, see feature::FeatureFileConfig.

operator: Option<bool>

Allow to lookup if operator is installed on cluster and use it

kubeconfig: Option<String>

Path to a kubeconfig file, if not specified, will use KUBECONFIG or ~/.kube/config or the in-cluster config.

sip_binaries: Option<VecOrSingle<String>>

Binaries to patch (macOS SIP). Use this when mirrord isn’t loaded to protected binaries that weren’t automatically patched. Runs endswith on the binary path (so bash would apply to any binary ending with bash while /usr/bin/bash would apply only for that binary).

Implementations

impl LayerFileConfig

pub fn from_path<P>(path: P) -> Result<Self, ConfigError>where P: AsRef<Path>,